fix: Catch exception on public pageé
